updated poolmarket contract

+++ contracts/IPoolMarket.sol
@@ -4,10 +4,10 @@ pragma solidity ^0.8.4;
 interface IPoolMarket {
   struct DispatchedOffer {
     address supplierAccount; //Offer supplier account
-    uint16 dispatchedAmount; //Dispatched amount
-    uint256 dispatchedAt; //Dispatched timestamp
+    uint dispatchedAmount; //Dispatched amount
+    uint dispatchedAt; //Dispatched timestamp
   }
 
-  function getPoolPrice(uint hour) external view returns(uint16);
+  function getPoolPrice(uint hour) external view returns(uint);
   function getDispatchedOffers(uint hour) external view returns(DispatchedOffer[] memory);
 }
